重构数据缓存的设计，现在的方式是通过反射结果的属性，根据属性的类型自动 生成对应类型的缓存池，确保和业务逻辑解耦。

+++ tools/SunEyeVision.Tool.Threshold/ThresholdResults.cs
@@ -25,7 +25,7 @@ namespace SunEyeVision.Tool.Threshold
         /// <summary>
         /// 最大值
         /// </summary>
-        public int MaxValueUsed { get; set; }
+        public double MaxValueUsed { get; set; }
 
         /// <summary>
         /// 使用的阈值类型
